feat: 에코딜 상품 목록 가져오는 api 구현

+++ src/main/java/org/phoenix/planet/dto/product/response/EcoProductListResponse.java
@@ -15,17 +15,17 @@ public class EcoProductListResponse {
 
     private long productId;
 
-    private long categoryId;
+    private Long categoryId;
 
     private String categoryName;
 
-    private long brandId;
+    private Long brandId;
 
     private String name;
 
     private int price;
 
-    private long quantity;
+    private int quantity;
 
     private int salePercent;
 }
